Make sure the check for game version in Makefile is only made when we're building, so we can clean without specifying it

+++ Makefile
@@ -14,13 +14,11 @@ INC_FLAGS := $(addprefix -I,$(INC_DIRS))
 CPPFLAGS ?= $(INC_FLAGS) -MMD -MP -O3 -march=native
 
 ifeq ($(game_version),1-1)
-CPPFLAGS += -D GAME_VERSION=0
+GAME_VERSION_ID := 0
 else ifeq ($(game_version),1-2)
-CPPFLAGS += -D GAME_VERSION=1
+GAME_VERSION_ID := 1
 else ifeq ($(game_version),2)
-CPPFLAGS += -D GAME_VERSION=2
-else 
-$(error Please specify a valid game_version, for instance "make game_version=1-2")
+GAME_VERSION_ID := 2
 endif
 
 CHECK_GAME_VERSION:=$(BUILD_DIR)/.last-game-version.$(game_version)
@@ -32,11 +30,13 @@ $(BUILD_DIR)/$(TARGET_EXEC): $(OBJS)
 		-lrt -Wl,--whole-archive -lpthread -Wl,--no-whole-archive  # Avoid segmentation fault when joining the threads. See https://stackoverflow.com/a/45271521
 
 $(BUILD_DIR)/%.cpp.o: %.cpp $(CHECK_GAME_VERSION)
+	@if [ -z ${GAME_VERSION_ID} ]; then >&2 echo Please specify a valid game_version, for instance \"make game_version=1-2\"; exit 1; fi
 	$(MKDIR_P) $(dir $@)
-	$(CXX) $(CPPFLAGS) $(CXXFLAGS) -c $< -o $@
+	$(CXX) $(CPPFLAGS) -D GAME_VERSION=$(GAME_VERSION_ID) $(CXXFLAGS) -c $< -o $@
 
 # https://stackoverflow.com/q/65400877/2761541
 $(BUILD_DIR)/.last-game-version.$(game_version):
+	@if [ -z ${game_version} ]; then >&2 echo Please specify a game_version, for instance \"make game_version=1-2\"; exit 1; fi
 	-rm -f $(BUILD_DIR)/.last-game-version.*
 	$(MKDIR_P) $(dir $@)
 	touch $@
